+++ src/main/kotlin/hu/infokristaly/forrasimageserver/repository/ClerkCRUDRepository.kt
@@ -3,5 +3,5 @@ package hu.infokristaly.forrasimageserver.repository
 import hu.infokristaly.forrasimageserver.entity.Clerk
 import org.springframework.data.repository.CrudRepository
 
-interface ClerkCRUDRepository : CrudRepository<Clerk, Int> {
+interface ClerkCRUDRepository : CrudRepository<Clerk, Long> {
 }
