+++ src/main/kotlin/hu/infokristaly/forrasimageserver/repository/OrganizationCRUDRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface OrganizationCRUDRepository : CrudRepository<Organization, Int> {
+interface OrganizationCRUDRepository : CrudRepository<Organization, Long> {
 }
